resplit governance tests
