+++ contracts/storage/QuadPassportStore.sol
@@ -1,32 +1,12 @@
 //SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.4;
 
-import "../interfaces/IQuadPassport.sol";
+import "../interfaces/IQuadPassportStore.sol";
 import "../interfaces/IQuadGovernance.sol";
 
 import "./QuadConstant.sol";
 
-contract QuadPassportStore is QuadConstant {
-    struct Attribute {
-        bytes32 value;
-        uint256 epoch;
-        address issuer;
-    }
-
-    /// @dev AttributeSetterConfig contains configuration for setting attributes for a Passport holder
-    /// @notice This struct is used to abstract setAttributes function parameters
-    /// `attrKeys` Array of keys defined by (wallet address/DID + data Type)
-    /// `attrValues` Array of attributes values
-    /// `tokenId` tokenId of the Passport
-    /// `issuedAt` epoch when the attribute has been attested by the Issuer
-    /// `fee` Fee (in Native token) to pay the Issuer
-    struct AttributeSetterConfig {
-        bytes32[] attrKeys;
-        bytes32[] attrValues;
-        uint256 tokenId;
-        uint256 issuedAt;
-        uint256 fee;
-    }
+contract QuadPassportStore is IQuadPassportStore, QuadConstant {
 
     IQuadGovernance public governance;
     address public pendingGovernance;
@@ -43,7 +23,7 @@ contract QuadPassportStore is QuadConstant {
     mapping(bytes32 => Attribute[]) internal _attributes;
 
     // Key could be:
-    // 1) keccak256(userAddress, keccak256(attrType), issuer)
-    // 1) keccak256(DID, keccak256(attrType), issuer)
+    // 1) keccak256(keccak256(userAddress, keccak256(attrType)), issuer)
+    // 1) keccak256(keccak256(DID, keccak256(attrType)), issuer)
     mapping(bytes32 => uint256) internal _position;
 }
