+++ contracts/storage/QuadReaderStore.sol
@@ -9,11 +9,4 @@ import "./QuadConstant.sol";
 contract QuadReaderStore is QuadConstant{
     IQuadGovernance public governance;
     IQuadPassport public passport;
-
-    struct ApplyFilterVars {
-        uint256 gaps;
-        uint256 delta;
-        uint256 filteredIndex;
-    }
-
 }
