+++ contracts/tests/QuadGovernanceUpgrade.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.4;
 
 import "../QuadGovernance.sol";
 
-contract QuadGovernanceV2 is QuadGovernance {
+contract QuadGovernanceUpgrade is QuadGovernance {
     function getPriceETHV2() external pure returns (uint) {
         return 1337;
     }
