+++ contracts/tests/QuadPassportUpgrade.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.4;
 
 import "../QuadPassport.sol";
 
-contract QuadPassportV2 is QuadPassport {
+contract QuadPassportUpgrade is QuadPassport {
     function foo() external pure returns(uint256){
         return 1337;
     }
