+++ contracts/tests/QuadReaderUpgrade.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.4;
 
 import "../QuadReader.sol";
 
-contract QuadReaderV2 is QuadReader {
+contract QuadReaderUpgrade is QuadReader {
     function foo() external pure returns(uint256){
         return 1337;
     }
